create Presets CRUD, other things refactored

+++ app/Http/Controllers/Api/CoursesController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CourseRequest;
-use App\Http\Resources\CourseResource;
+use App\Http\Resources\CoursesResource;
 use App\Models\Course;
 use App\Repositories\CoursesRepository;
 use Illuminate\Http\Request;
@@ -15,33 +15,40 @@ class CoursesController extends Controller
 
     public function __construct(CoursesRepository $repository)
     {
-        $this->authorizeResource(Course::class);
         $this->repository = $repository;
     }
 
     public function index(Request $request)
     {
+        $this->authorize(Course::class);
+
         $courses = $this->repository->index($request);
 
-        return CourseResource::collection($courses);
+        return CoursesResource::collection($courses);
     }
 
-    public function show(Course $course, Request $request)
+    public function show(Request $request)
     {
-        return new CourseResource($course);
+        $this->authorize(Course::class);
+
+        $course = $this->repository->show($request);
+
+        return new CoursesResource($course);
     }
 
     public function store(CourseRequest $request)
     {
-        $course = $request->user()->isAdmin() ?
-            Course::create($request->validated()) :
-            $request->user()->courses()->create($request->validated());
+        $this->authorize(Course::class);
+
+        $course = $this->repository->store($request);
 
         return redirect()->route('courses.show', compact('course'));
     }
 
     public function update(CourseRequest $request, Course $course)
     {
+        $this->authorize($course);
+
         $course->update($request->validated());
 
         return redirect()->route('courses.show', compact('course'));
@@ -49,6 +56,8 @@ class CoursesController extends Controller
 
     public function destroy(Course $course)
     {
+        $this->authorize($course);
+
         $course->delete();
 
         return redirect()->route('courses.index');
